Refactored LoginPage and Step Definitions: Integrated CommonActions and CommonAsserts in LoginPage. Updated ByControlWrapper instantiation with controlName parameter. Modified Step Definitions to align with the AssertLoggedIn method.Also CommonFunction merged with CommonAction class
