+++ bin/Debug/net6.0/TestExecution.json
@@ -1,1 +1,1 @@
-{"Nodes":[],"ExecutionTime":"2023-12-10T20:23:18.3728489Z","GenerationTime":"0001-01-01T00:00:00","PluginUserSpecFlowId":"7f2adf6f-73f4-46dd-8c55-3add25d1bdb4","CLIUserSpecFlowId":null,"ExecutionResults":[{"ContextType":"Scenario","FeatureFolderPath":"Features","FeatureTitle":"Login Functionalities","ScenarioTitle":"Valid Admin login","ScenarioArguments":[],"Status":"OK","StepResults":[{"Duration":"00:00:01.1697702","Status":"OK","Error":null,"Outputs":null},{"Duration":"00:00:00.2192987","Status":"OK","Error":null,"Outputs":null},{"Duration":"00:00:01.5550993","Status":"OK","Error":null,"Outputs":null},{"Duration":"00:00:00.0428025","Status":"OK","Error":null,"Outputs":null}],"Outputs":null}],"StepReports":null}
+{"Nodes":[],"ExecutionTime":"2023-12-10T22:24:25.5532495Z","GenerationTime":"0001-01-01T00:00:00","PluginUserSpecFlowId":"7f2adf6f-73f4-46dd-8c55-3add25d1bdb4","CLIUserSpecFlowId":null,"ExecutionResults":[{"ContextType":"Scenario","FeatureFolderPath":"Features","FeatureTitle":"Login Functionalities","ScenarioTitle":"Valid Admin login","ScenarioArguments":[],"Status":"OK","StepResults":[{"Duration":"00:00:03.9755183","Status":"OK","Error":null,"Outputs":null},{"Duration":"00:00:00.2543104","Status":"OK","Error":null,"Outputs":null},{"Duration":"00:00:01.6274855","Status":"OK","Error":null,"Outputs":null},{"Duration":"00:00:00.0229882","Status":"OK","Error":null,"Outputs":null}],"Outputs":null}],"StepReports":null}
